Pull common extensions to the `psi` module
Also:
  * Evaluate `documentManager` once.
